+++ lib/routing/app_routes.dart
@@ -5,4 +5,6 @@ class AppRoutes {
 
   static const home = '/';
   static const settings = '/settings';
+  static const clients = '/clients';
+  static const clientNew = '/clients/new';
 }